requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
aiofiles>=23.0.0
groq>=0.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0